refactor : enum 클래스로 이메일 폼 위치, 제목 관리
