+++ src/main/java/koreatech/in/dto/normal/mail/Mail.java
@@ -1,4 +1,4 @@
-package koreatech.in.dto;
+package koreatech.in.dto.normal.mail;
 
 import lombok.Builder;
